mofified result as float values upto two digits

+++ main.py
@@ -31,8 +31,8 @@ def multiplication():
 def division():
     value1=request.args.get('A',default = 0, type = int)
     value2=request.args.get('B',default = 0, type = int)
-    result=value1/value2
-    return '%d \n' % result
+    result=float(value1/value2)
+    return '%2f \n' % result
 
 if __name__ == "__main__":
     app.run()
